Add files via upload
Defining Range DR and Implied Defining Range IDR for regular Session and overnight Session  This script is showing the IDR and DR for the regular trading session and for the overnight session based on the rules from the creator of the DR / IDR concept. It works for all major Forex Pairs, BTC , ETH and the US Equity indices. This concept is based on rules and has a 80 % probability to be correct.  It should be applied in the 5 Min. Timeframe.  The timings for the RDR are from 09.30 - 10.30 am New York local time. The timings for the ODR are from 03.00 - 04.00 am New York local time.  Rules: 1. If price in the 5 Min timeframe closes above the DR high after 10.30 am or 04.00 am then the DR low will be with 80 percent probability the low of the trading session. This is called confirmation. 2. If price in the 5 Min timeframe closes below the DR low after 10.30 am or 04.00 am then the DR high will be with 80 percent probability the high of the trading session. This is called confirmation. 3. If price closes above the IDR high after 10.30 am or 04.00 am it is an early indication that the low of the DR will be the low of the day and vice versa.